Desktop/Scorpius-X/backend/requirements_mev.txt
# Ultimate MEV Bot Dependencies
# Core web framework
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# Blockchain and Web3
web3==6.11.3
eth-account==0.9.0
eth-abi==4.2.1
eth-utils==2.3.1

# DeFi and MEV specific
flashbots==1.2.0
aiohttp==3.9.1
requests==2.31.0

# Data processing and ML
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.0

# Async and concurrency
asyncio-pool==0.6.0
aiofiles==23.2.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: Database support
redis==5.0.1
pymongo==4.6.0
